+++ .github/workflows/main.yml
@@ -13,9 +13,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install numpy
-          pip install pandas
-          #pip install -r requirements.txt
+          pip install -r requirements.txt
 
       - name: Run tests
         run: python -m unittest discover -s tests/
