+++ requirements.txt
@@ -1,3 +1,5 @@
 lifelines
 autogluon.tabular
 pytest
+numpy
+pandas
