+++ taimaka_health_predictions/utils/globals.py
